pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gene-classifier-ml"
version = "0.1.0"
description = "A machine learning pipeline for binary classification of gene functions"
readme = "Readme.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Dheeraj Ram", email = "dheerajram13@users.noreply.github.com"}
]
keywords = ["machine-learning", "bioinformatics", "gene-classification", "ensemble-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "pandas>=2.2.3",
    "numpy>=2.0.2",
    "scikit-learn>=1.5.2",
    "scipy>=1.13.1",
    "joblib>=1.4.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=24.10.0",
    "isort>=5.13.2",
    "pylint>=3.3.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
experiment = [
    "mlflow>=2.10.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]
dvc = [
    "dvc>=3.0.0",
    "dvc-s3>=3.0.0",
]

[project.scripts]
gene-classifier = "main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["gene_classifier*"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=.",
    "--cov-report=html",
    "--cov-report=term-missing",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pylint.messages_control]
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "R0913",  # too-many-arguments
    "R0914",  # too-many-locals
]

[tool.pylint.format]
max-line-length = 88
